+++ bin/remap-snapshot.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require('../lib/remap-snapshot-cli')(process.argv.slice(2));