+++ src/index.ts
@@ -1,3 +1,4 @@
 export {setup} from './setup';
 export {transform} from './transform';
 export {transform_actual} from './transform-actual';
+export {remap_snapshot} from './remap-snapshot';