+++ src/remap-snapshot.ts
@@ -1,0 +1,44 @@
+import * as ts from 'typescript';
+import {Trigger} from './definitions';
+import {create_triggers} from './utils/create-triggers';
+import {default_to} from './utils/default-to';
+import {get_formatted_description} from './utils/get-formatted-description';
+
+// tslint:disable-next-line:no-require-imports no-var-requires
+const require_from_string = require('require-from-string');
+
+export const remap_snapshot = (snapshot_content: string, source_content: string) => {
+  const snapshot_data: Record<string, string> = require_from_string(snapshot_content);
+
+  const source_file = ts.createSourceFile('', source_content, ts.ScriptTarget.Latest, false);
+  const triggers = create_triggers(source_file);
+
+  const source_content_lines = source_content.split('\n');
+
+  const counters: Record<string, number> = {};
+  triggers.forEach(trigger => {
+    const title = get_snapshot_title(trigger);
+    const counter = counters[title] = default_to<number>(counters[title], 0) + 1;
+
+    const description = `${title} ${counter}`;
+    if (description in snapshot_data) {
+      const {line} = trigger;
+      let snapshot = snapshot_data[description].trim();
+      const breakline = snapshot.indexOf('\n');
+      if (breakline !== -1) {
+        snapshot = `"${snapshot.slice(1, breakline)}"`;
+      }
+      source_content_lines[line] += ` -> ${JSON.parse(snapshot)}`;
+    }
+  });
+
+  return source_content_lines.join('\n');
+};
+
+function get_snapshot_title(trigger: Trigger) {
+  let title = get_formatted_description(trigger, false);
+  if (trigger.group !== undefined) {
+    title = `${trigger.group.title} ${title}`;
+  }
+  return title;
+}